backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
tensorflow==2.12.0
pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0